Add igw utils module to ensure no closures
